Program currently finds and connects to the Disto D810.
Another activity has been added that is used as a Dialog
interface.

The program currently scans, finds, and stores Remote Le
Devices. As each device is found, the device name is presented
to the user in a ListView as an item. When the user clicks
on a ListView item (the device name), the device represented
by this item is connected to.

Program does not currently interact with Disto device
beyond connecting.

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>