Tablet connects to Disto and turns on the laser, but is very unreliable.
Unnecessary code has been removed.

The BluetoothGattCallback object -- the one previously created and defined
in the BluetoothLeService -- has been turned into a new class
(GattCallback) that extends the BluetoothGattCallback class. Now instead
of creating an instance of BluetoothGattCallback and overriding some the
class's functions upon instantiation, the functions are overriden in the
GattCallback class; an instance of GattCallback replaces the old instance
of BluetoothGattCallback.

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>