Application now connects to Disto and reads and displays a distance value.
The distance value is read when it changes on the Disto. The app subscribes
to the Disto's distance characteristic to be notified when the distance
does change. Once the distance is read, the value is added to the table in
the MainActivity.

The value is read in meters; it should be displayed in inches and feet.

Connecting to the device is still unreliable.

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>