Users can now edit a row by clicking on it.
When in the TableRowEditorActivity, if the "renumber all below" checkbox
is checked, all rows below the row being edited are renumbered according
to the Pipe Number of the edited row.

The app is in the process of becoming truly fullscreen.

The Job Info button launches an activity, displaying the user with
editable information about the job at hand. The activity does not currently
store any of the information obtained from this activity.

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>