The app is now fullscreen for all activities.
Added code to each activity: added to the onResume and code that listens
for UI changes and window focus changes to each activity. When the activity
has focus or the system bars are showing, or the app is resumed, the system
UI visibility is set to specific UI options. This ensures that the app
never leaves fullscreen mode.

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>