Users can now delete, close, and create jobs from the Job Info menu.
An activity was created for the Job Info menu (JobInfoMenuActivity).
It has five buttons with different functionalities:
	open existing job
	create a new job
	close this job
	delete this job

Pressing the open existing job button launches OpenJobActivity. The
activity currently gets all of the jobs from the jobsDir and displays them
to the user. Clicking on a job doesn't open the job yet.

Pressing the create a new job button launches the CreateJobActivity.

Pressing the close this job button takes the user back to the MainActivity.

Pressing the delete this job button deletes the current job and takes the
user back to the MainActivity.

The open existing job in the MainActivity has the same functionality as
the open existing job in the JobInfoMenuActivity.

CreateJobActivity and JobInfoActivity now check to see if the job name
entered by the user already exists.

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>