Created an activity to be used for verifying actions.
The VerifyActionActivity is used to verify actions. Text to be displayed
in the main TextView of the activity can be sent to the activity using
intent extras.

When the delete this job button is pressed, the VerifyActionActivity is
launched to verify that the user really wants to delete the job. Cancel
works but the JobInfoMenuActivity does not currently listen for results;
therefore, clicking the I'm sure button currently has no effect except for
closing the VerifyActionActivity.

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>