Started making tally device connection code generic.
Created TallyDeviceService class to replace BluetoothLeService class.

The new class uses functions within TallyDeviceConnectionHandler for all
connection operations.

The new concept uses children of the TallyDeviceConnectionHandler class
which override functions in the parent class. All connection functions in
the parent class are kept generic so that they may be used with any
connection type (BluetoothLe, simulation, etc.).

TallyDeviceService creates a variable of type TallyDeviceConnectionHandler
and assigns the variable to be a pointer to an instance of one of
TallyDeviceConnectionHandler's children, depending on the connection type.

For example:
TallyDeviceConnectionHandler handler = new TallyDeviceBluetoothLeHandler();

With this idea, switching between connection types becomes a simple matter
of changing the class type of the instance the handler variable was
assigned to. After the class type has been changed, all of the same
function calls can still be used without having to change any code in the
TallyDeviceService.

For example:
TallyDeviceConnectionHandler handler = new TallyDeviceBluetoothLeHandler();
	can be changed to
TallyDeviceConnectionHandler handler = new TallyDeviceSimulationHandler();

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>