Removed the need of the MSG_CONNECTED_TALLY_DEVICE_NAME.
If the connection state is CONNECTED or CONNECTING, the tally device name
that the device is interacting with is sent in the connection state message
from the TallyDeviceService to its messenger client.

Removed the MSG_CONNECTED_TALLY_DEVICE_NAME handling code from the
TallyDeviceConnectionStatusMessageActvitiy class.

Removed any code to do with MSG_CONNECTED_TALLY_DEVICE_NAME from the
TallyDeviceService class.

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>