Changed JobDisplayActivity.java to be compatible with TallyDeviceService.
All code that had anything to do with Bluetooth was removed from
JobDisplayActivity.

The IncomingHandler class inside of JobDisplayActivity was changed to
accept messages from the TallyDeviceService.

Some new handling functions were added to JobDisplayActivity. All of the
functions were put in alphabetical order.
All logs were removed.

The adjusted column is now changed when the total length of the row in the
TableRowEditorActivity is changed.

Some refactoring and comment changes were done.

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>