Solved the context null pointer exception.
The context used in TallyDeviceBluetoothLeConnectionHandler was giving
a null pointer exception because the context passed to it upon
instantiation was null and was never set after that. Now when an activity
registers with TallyDeviceService, the context of the activity is sent to
the service in the registering message. The service then passes the context
on to the TallyDeviceConnectionHandler, thus solving the context null
pointer exception.

All code for sending the register message to the TallyDeviceService has
been moved out of the overridden functions in the ServiceConnection
variable and into a single separate function (registerWithService()). This
has been done for all activities.

Exiting the TallyDeviceScanActivity now stops the scan if one is in
process.

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>