Connecting to the tally device now works 100% of the time.
A workaround was implented to solve the problem of the accidental
disconnect from the tally device. To solve the issue, the Bluetooth of the
tablet was turned off and on programmatically and then the tally device
would be scanned for, found, and reconnected to.

The Bluetooth had to be turned off and on because if the device was
reconnected to without restarting the Bluetooth, writing to a descpriptor
resulted in a status of GATT_INSUFFICIENT_ENCRYPTION. Restarting the
Bluetooth solved this issue.

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>