Added code to handle the tally device name being found null in the BLE scan.
For a reason currently unknown, the tally device name found by the BLE scan
can be null. To compensate for this issue, the scan is stopped and
restarted  each time a null device name is passed to onLeScan(). This will
continue to happen for the intended duration of the BLE scan; once the
time specified for the scan ends, the scan will no longer be restarted
because of a null device name.

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>