The GATT_INSUFFICIENT_ENCRYPTION descriptor write error is now solved by unbonding from the tally device if it is bonded.
The descriptor write error turned out to be a pairing error. When the
discoverServices() function was called, sometimes the tally device was
unpaired and disconnected from the tablet. But, when the tally device was
reconnected to (without restarting the Bluetooth on the tablet), the
Android OS still seemed to think that it was still paired to the tally
device. Because the OS thought it was still paired to the tally device,
it did not go through the usual bonding process and instead gave a
descriptor write status of GATT_INSUFFICIENT_ENCRYPTION. The OS only
realized that it was no longer paired with the tally device, when the
Bluetooth was restarted; the descriptor write status of
GATT_INSUFFICIENT_ENCRYPTION would reoccur every time a descriptor was
written to until the Bluetooth on the Nexus was restarted.

Before this check in, the GATT_INSUFFICIENT_ENCRYPTION descriptor write
error that was solved by resetting the Bluetooth on the Nexus 7 after an
unsuccessful attempt to connect to the tally device. After the Bluetooth
was restarted, the tally device was reconnected to.
(See previous Git check ins)

Now, the error is solved by unpairing from the tally device before the
BLE scan is started.

When the startScanForTallyDevices() function is called, a list of bonded
devices is retrieved from the BluetoothAdapter. If the tally device is one
of the paired devices, the tally device is unpaired programmatically.
Then, the BLE scan is started.
Removing the bond (pairing) prevents the risk of unbonding from the tally
device without the Android OS realizing it; by initiating the unpairing
process programmatically, the OS is made fully aware.

The new solution is better because it takes less time and does not turn
off the Bluetooth on the Nexus, which has the potential of disconnecting
other devices besides the tally devices.

Code is still sloppy and needs to be cleaned up; old method of solving
error needs to be removed.

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>