Totals of columns and scroll to bottom of table abilities added.
The totals of the Total Length and Adjusted columns are summed up and
inserted into a table below the MeasurementsTable each time a new distance
value is received or the redo button is pressed.

Added code to send the connection state to the
TallyDeviceConnectionStatusMessageActivity when it registers to the
service.

Signed-off-by: Hunter Schoonover <hunter@folsomvillage.com>