Fixed JobDisplayActivity crashing when a new job was created bug.
The bug was solved by adding the JOB_INFO_INCLUDED key was added to the
intent extras that starts the JobDisplayActivity so that JobDisplayActivity
knows to pull the job info from the intent extras. This was a problem
before because the jobInfo variable was never being set, so
JobDisplayActivity was passing around and attempting to use a null object.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>