Cal value edit text field now set and the spinner for add or sub is gone.
The calibration value edit text field in the MoreOptionsActivity is now set
to the proper value stored in SharedSettings.

If the calibration value should be subtracted, then the user can now add a
negative sign to the edit text field instead of having to change the
spinner from add to subtract (the spinner has been removed).

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>