Necessary values are now converted between Imperial and Metric.
TallyGoal and the calibation value are now converted between Imperial and
Metric. Both their Imperial and their Metric values are saved to file so
that they only have to be converted once (for accuracy).

The Imperial and the Metric values for minimum and maximum values allowed
are now stored to file.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>