Started making the app work on both the tablet and the phone.
Added universal dimensions to dimens.xml and integrated them throughout the
layout files.

When a large screen is detected, the "values-large/dimens.xml" file is
automatically used.

When a normal screen is detected, the "values-normal/dimens.xml" file is
automatically used.

Right now, both of the dimens files contain the same values (for the large
screen).

TODO: Change the dimentions in "values-nomral/dimens.xml" so that they
	accomodate for a normal screen.

Also changed some LinearLayout layouts to RelativeLayout layouts for
universability, flexibity, and easier programming. This was done to all
dialog actities containing a title bar:
	activity_job_info.xml
	activity_job_info_menu.xml
	activity_more.xml
	activity_more_options.xml
	activity_open_job.xml
	activity_table_row_editor.xml
	activity_tally_device_scan.xml
	activity_verify_action.xml

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>