Arrow keys now work for the Job Display Activity.
Nothing in activity_job_display.xml is focusable except for the TableRows
programmatically added to the measurements table. This gives us full
control of the focus order.

Measuring and pressing the redo button now focuses the last row in the
measurements table.

The background color of a selected row is no longer changed back to its
original color after the TableRowEditor activity ends. It is now
programmatically brought into and left in focus so that the user can
easily see which row was edited last.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>