TableRowEditorActivity no longer crashes app when device is put to sleep.
The Parcelable.CREATOR variables are no longer initialized in functions,
they are now final and intialized in the class. Not being final and not
being initialized immediately upon class creation caused problems when it
came to the parcelable aspect of the two classes.

StandardActivity::Requesting focus of a view is now encased in a try block.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>