Bad sound no longer plays randomly when there was no bad measurement received.
The bug was caused because the timers posted to the handler were
overlapping each other so that the boolean wasn't set properly.

Now, the noMeasurementReceived runnable is posted to the handler with a
one second delay when the measure command is sent to the tally device, and
when the measurement value is received the runnable is removed from the
handler.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>